Switch to using environment vars for configuration (#22)
While flags are certainly better for applications that the user will be
running in their own environment, sos-journaler will probably always be
run in Docker through Docker Compose. Environment variables are better
supported with Docker Compose and doing so allows us to switch from
using ENTRYPOINT to CMD. CMD is a better match for what we're trying to
accomplish and seems to interact better with tooling.